+++ app/build.gradle
@@ -1,5 +1,5 @@
-import com.example.marvelapp.Dependencies
-import com.example.marvelapp.Version
+import com.example.buildsrc.Dependencies
+import com.example.buildsrc.Version
 
 plugins {
     id 'com.android.application'
