+++ build.gradle
@@ -1,4 +1,4 @@
-import com.example.marvelapp.Dependencies
+import com.example.buildsrc.Dependencies
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
